requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
xgboost==2.0.3
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
networkx==3.2.1
joblib==1.3.2
streamlit==1.31.0
plotly==5.18.0
prometheus-client==0.19.0
python-dotenv==1.0.0
httpx==0.26.0
faker==22.2.0
